[REF] getters: move core getters to getters file
Task: 6428159
