+++ src/plugins/core/unsquisher.ts
@@ -5,8 +5,8 @@ import { toCartesian } from "../../helpers/coordinates";
 import { expandRange, expandXc } from "../../helpers/expand_range";
 import { formatValue } from "../../helpers/format/format";
 import { CoreCommand, UpdateCellCommand } from "../../types/commands";
-import { CoreGetters } from "../../types/core_getters";
 import { Format } from "../../types/format";
+import { CoreGetters } from "../../types/getters";
 import { Position, UID } from "../../types/misc";
 import { Range } from "../../types/range";
 import { NO_CHANGE, SEPARATOR, SquishedContent, SquishedFormula } from "./squisher";
