+++ src/registries/chart_data_source_registry.ts
@@ -9,8 +9,7 @@ import {
   ExcelChartDefinition,
 } from "../types/chart/chart";
 import { CommandResult } from "../types/commands";
-import { CoreGetters } from "../types/core_getters";
-import { Getters } from "../types/getters";
+import { CoreGetters, Getters } from "../types/getters";
 import { RangeAdapterFunctions, UID } from "../types/misc";
 import { Validator } from "../types/validator";
 import { Registry } from "./registry";
